chore: Extracted class for better maintainability
