+++ koswat/cost_report/io/summary/summary_locations/cluster_geodataframe_output_fom.py
@@ -1,0 +1,37 @@
+from geopandas import GeoDataFrame
+
+
+class ClusterGeoDataFrameOutputFom:
+    """
+    Data structure to wrap the output generated by `ClusterCollectionShpFom.generate_geodataframes`.
+    """
+
+    base_layer: GeoDataFrame | None
+    initial_state: GeoDataFrame | None
+    new_state: GeoDataFrame | None
+
+    def __init__(self, gdf_tuples: tuple[GeoDataFrame]):
+        self.base_layer = None
+        self.initial_state = None
+        self.new_state = None
+
+        if not gdf_tuples:
+            return
+        self.base_layer = gdf_tuples[0]
+        self.initial_state = gdf_tuples[1]
+        self.new_state = gdf_tuples[2]
+
+    def is_valid(self) -> bool:
+        """
+        Checks that all `GeoDataFrame` properties are set.
+
+        Returns:
+            bool: Validation result.
+        """
+
+        def validate_gdf(gdf: GeoDataFrame | None) -> bool:
+            return gdf is not None
+
+        return all(
+            map(validate_gdf, [self.base_layer, self.initial_state, self.new_state])
+        )
